Resolved comments from first submission
